MAJOR CHANGE - PLEASE READ DETAILS: De-coupling of styles that were too specific and didn't provide enough flexibility
After some work with the template we have found that some styles for Joomla and some extensions that we created initially have proven to be too specific, and we have been forced to remove on projects. This is contrary to the spirit of the template. For that reason we have removed them from the main LESS structure. We moved the FLEXIcontent  and RegistrationPro rules to a new folder "library" that will contain suggested CSS styles that you can use in your project only if you think they are needed or useful. The Joomla menu styles are removed completely and we will be developing new classes for implementing a more robust menu system in 06-common-menu.less.

You may need to go back to the following files and add them to your custom.less:

08-joomla-menu.less - https://github.com/Lyquix/tpl_lyquix/blob/7707976efb23224dd0ac3d00e261d08039c7ab4c/css/less/08-joomla-menu.less
08-joomla-registrationpro.less - https://github.com/Lyquix/tpl_lyquix/blob/948e4b07af895c2b7c80c4465fa295ec30a8a0d4/css/less/08-joomla-registrationpro.less
09-flexicontent.less - https://github.com/Lyquix/tpl_lyquix/blob/891d8e903ee84859e20e3febd6643d76d8f26fe2/css/less/09-flexicontent.less